moved LoggerWithoutDateInOutput class
